+++ examples/04_3D/src/Example3D.cpp
@@ -12,8 +12,8 @@ Example3D::Example3D(iWindowPtr window)
 void Example3D::onInit()
 {
     // setup camera
-    // we want a camera which can be rotated arround the origin
-    // we will acchive that with 3 transform nodes
+    // we want a camera which can be rotated around the origin
+    // we will achieve that with 3 transform nodes
     // one is for the heading
     iNodeTransform *cameraHeading = iNodeManager::getInstance().createNode<iNodeTransform>();
     // give the transform node a name. naming is optional and ist jus for helping to debug.