+++ src/igor/igor/resources/texture/iTextureFont.cpp
@@ -26,14 +26,15 @@ namespace igor
         const iaString resolved = iResourceManager::getInstance().getPath(filename);
 
         iParameters param({{"name", resolved},
+                           {"type", iaString("texture")},
                            {"buildMode", iTextureBuildMode::Normal}});
-        _texture = std::dynamic_pointer_cast<iTexture>(iResourceManager::getInstance().loadResource(param));
+        _texture = iResourceManager::getInstance().loadResource<iTexture>(param);
 
         if (!_texture->isValid())
         {
             return;
         }
-        _pixmap = iTextureFactory::loadPixmap(resolved);
+        _pixmap = iPixmap::loadPixmap(resolved);
 
         if (_pixmap == nullptr)
         {